1. 加入流水号 2. 备份数据时空值判断

+++ lmrj-modules/lmrj-module-dsk/src/main/java/com/lmrj/dsk/eqplog/service/impl/EdcDskLogProductionServiceImpl.java
@@ -88,7 +88,9 @@ public class EdcDskLogProductionServiceImpl  extends CommonServiceImpl<EdcDskLog
                 }
             }
         }
-        super.deleteBatchIds(deleteList);
+        if(deleteList.size() != 0 ){
+            super.deleteBatchIds(deleteList);
+        }
         return hisList;
     }
